requirements.txt
telethon
python-dotenv
flask
aiohttp
gunicorn